Factor out V extension macros into their own header
No functional change.
